Update to allow F-I curve generation
git-svn-id: svn://neuroconstruct.org:3999/neuroConstruct/trunk@1179 cb8a3d95-5132-0410-b650-8d485996f008

+++ src/ucl/physiol/neuroconstruct/simulation/SimulationData.java
@@ -1399,7 +1399,7 @@ public class SimulationData
             if (pso!=null) synInfo =  " (synapse: "+pso.getSynRef()+")";
             String info = "DataStore"+synInfo+" for "+variable+" on segment: "+segId+" on "+ getCellRef();
 
-            info = info + ": ("+dataPoints[0]+", "+dataPoints[1]+", ... ["+dataPoints.length+" entries])";
+            info = info + ": ("+(float)dataPoints[0]+", "+(float)dataPoints[1]+", ... ["+dataPoints.length+" entries])";
 
             return info;
         }
